refactor: change to searching @ComponentScan first in ClassBeanScanner
