+++ nextstep-di/src/test/java/nextstep/di/factory/example/component/BeanInjectedComponent.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.factory.example.component;
 
 import nextstep.annotation.Inject;
 import nextstep.stereotype.Controller;
